:tada: Initial commit
There's very little here right now. This commit is really about getting a
basic development environment framework up and running for the project, with
the most minimal of code to run a couple of tests against. From here on in I
want to be making fairly atomic commits as the library grows.

It's how I prefer to do things, and it also means I always have a pretty
good backup copy of the repo off my machine.