:sparkles: Add some of the basic Norton Guide IO methods
Core here is the method to "decrypt" the data inside the database, and to
then read a byte and optionally apply the decryption. Then, on top of this,
add a method for reading a word. Reading a long will come soon, when I need
that.

Also add a method for skipping a number of bytes.

It's worth noting that the NortonGuide object is expecting to keep the state
of the position as the actual position in the file handle.