:sparkles: Get line loading going for long entries
Still need to add see-also entries, and there's a TODO for this. Right now
this also means the code still isn't in a state where I can load one entry
after another and be sure that the file position is placed over the next
entry. That's fine, that's on purpose, I'm just pecking away at this, a
little bit each evening.