:white_check_mark: Expand the navigation tests to start testing EOF
In doing this, I've uncovered a bug when reading a non-trivial guide. I'm
getting this sort of thing:

>>> ng.load()
Traceback (most recent call last):
  File "/Users/davep/develop/python/ngdb.py/ngdb/entry.py", line 146, in load
    return cls._map[ EntryType( guide.peek_word() ) ]( guide )
  File "/Users/davep/.pyenv/versions/3.10.0/lib/python3.10/enum.py", line 384, in __call__
    return cls.__new__(cls, value)
  File "/Users/davep/.pyenv/versions/3.10.0/lib/python3.10/enum.py", line 701, in __new__
    raise ve_exc
ValueError: 256 is not a valid EntryType

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/davep/develop/python/ngdb.py/ngdb/guide.py", line 47, in _guard
    return meth( self, *args, **kwargs )
  File "/Users/davep/develop/python/ngdb.py/ngdb/guide.py", line 219, in load
    return Entry.load( self._guide )
  File "/Users/davep/develop/python/ngdb.py/ngdb/entry.py", line 148, in load
    raise UnknownEntryType( f"Unknown guide entry type: {guide.peek_word()}" ) from error
ngdb.types.UnknownEntryType: Unknown guide entry type: 256

This is just a couple of entries into the guide (in this case oslig.ng).
Seems pretty obvious that somewhere in the entry reading code I'm ending up
a byte or two off (most likely), so I just need to hunt through the entry
code and be sure it's properly correct.

For now though... here's the tests that will help test that.