:fire: Remove all attempts at CI on GitHub for now
It looks handy, but I'm a) too used to GitLab and b) my normal approach to
developing Python code, and my use of Pipenv, doesn't seem to mix well with
it (especially when testing against different versions of Python).

I *like* how it works, but it's going to take some more thought, so I'm not
going to get distracted now.