:green_heart: Second stab at getting actions going
I want to lint, type check and unit test the library in as many Python 3.x
versions as possible. Meanwhile, I tend to use pipenv to manage my
day-to-day coding within the library repo. This doesn't really play well
with pipenv as it (rightly?) expects the Python version to not be pulled
from under it.

So this time around I do some nasty patching to the Pipfile to make pipenv
think that we were always using a particular version of Python.

Yeah, I know...