:construction: Kick off the base of the parser
This isn't the finished product; this is pretty much the core of jsNG parser
turned into Python. This isn't the final form by any stretch. There's way
too much branching going on here and far too much code inside a constructor.
But this gives me a starting point.

From here I want to refactor the whole thing so it's far more Pythonic. But
it gives me a starting point.