:hammer: More experimenting with the base parser
Here I drop using a dataclass for the class that helps track state and make
it into a full class and give it a bit more work to do. I also then really
pare back the base class' constructor.

Again, I've not tested this yet, not run this code. That's a feature, not a
bug. I'm stills settling on a design for this bit of code that feels good in
Python.