:white_check_mark: Start some parser tests
In this case start throwing some simple tests at the plain text parser.
There's a lot of work to be doing before I can call the parser properly
tested -- I should really create a custom unit testing parser so that all
the different states can be tested. But that's later.