:bookmark: v0.1.0
Not too much has changed, but I'm now at a point where I want to try and
start work on the Python take on ng2html, and I want to build that as a
different library that has this as a requirement. So time to draw a line on
this and start to look at that.

Working on ng2html will help me test what I've got going here so far, for
sure.