:hammer: Make it clear that some parameters really are unused
Depending on the tool I'm using to test/check/lint things, I get complaints
about some of the arguments for some of the methods in BaseParser being
unused. Thing is... they're supposed to be unused and those methods are
supposed to be no-op methods. On the other hand they're not abstract
methods, far from it. The idea is that someone can inherit from BaseParser
and either decide to implement a specific method or not, and not doing so is
cool.

So... `del arg` is a totally acceptable way of consuming the argument
without actually doing anything with it.