:sparkles: Add experimental markup parsing
In the first instance I'm using this to add a parser that turns the guide
text into Rich console markup:

  https://rich.readthedocs.io/en/stable/protocol.html

But the base class for that should also serve as the base class for the HTML
parser in ng2web (which I'll need to rewrite once this is done).

This isn't 100% working yet because I need to sort out the colour mapping --
I'd forgotten that the MS-DOS colour values don't match what you get in your
average Unix console.

So this is more or less the main version of this code, and to come are some
tweaks to get the colours mapping correctly.

See
https://rich.readthedocs.io/en/stable/appendix/colors.html#appendix-colors
to get me going.