:sparkles: Add a tool for testing all the guides I have
In adding a global search facility to AgiNG:

  https://github.com/davep/aging/pull/18

I encountered a small number of guides that have unexpected or malformed
content; this in turn caused crashes. Having run into this sort of thing
with WEG, and added various workarounds for these problems, I decided it
would make sense to have a tool in this repository that would mimic the
global search.

So this tool reads and parses every single entry in every single guide I
have available, or a specific guide if I want to do a quick test, and it
reports any errors it finds.

Once this tool emits zero errors, thus taking the approach of accepting
broken guides so make them work, the whole library will be even more robust.